Moved rules to Rule.hs
