+++ lib/theory/src/OpenTheory.hs
@@ -6,7 +6,8 @@
 
 module OpenTheory (
     module OpenTheory
-    ,module SpecificProof
+    , module SpecificProof
+    , module Rule
     ) where
 
 import GHC.Generics
@@ -43,232 +44,9 @@ import Theories
 
 import SpecificProof
 
-
--- Lemma construction/modification
-----------------------------------
-
--- | Create a new unproven lemma from a formula modulo E.
-unprovenLemma :: String -> [LemmaAttribute] -> TraceQuantifier -> LNFormula
-              -> Lemma ProofSkeleton
-unprovenLemma name atts qua fm = Lemma name qua fm atts (unproven ())
-
-skeletonLemma :: String -> [LemmaAttribute] -> TraceQuantifier -> f -> p -> ProtoLemma f p
-skeletonLemma name atts qua fm = Lemma name qua fm atts
-
--- | Create a new unproven diff lemma.
-unprovenDiffLemma :: String -> [LemmaAttribute]
-              -> DiffLemma DiffProofSkeleton
-unprovenDiffLemma name atts = DiffLemma name atts (diffUnproven ())
-
-skeletonDiffLemma :: String -> [LemmaAttribute] -> DiffProofSkeleton -> DiffLemma DiffProofSkeleton
-skeletonDiffLemma name atts = DiffLemma name atts
+import Rule
 
 
-
-
-
-
-------------------------------------------------------------------------------
--- Commented sets of rewriting rules
-------------------------------------------------------------------------------
-
--- | A protocol rewriting rule modulo E together with its possible assertion
--- soundness proof.
--- Optionally, the variant(s) modulo AC can be present if they were loaded
--- or contain additional actions.
-data OpenProtoRule = OpenProtoRule
-       { _oprRuleE  :: ProtoRuleE             -- original rule modulo E
-       , _oprRuleAC :: [ProtoRuleAC]          -- variant(s) modulo AC
-       }
-       deriving( Eq, Ord, Show, Generic, NFData, Binary )
-
--- | A diff protocol rewriting rule modulo E
--- Optionally, the left and right rules can be present if they were loaded
--- or contain additional actions.
-data DiffProtoRule = DiffProtoRule
-       { _dprRule       :: ProtoRuleE         -- original rule with diff
-       , _dprLeftRight  :: Maybe (OpenProtoRule, OpenProtoRule)
-                                              -- left and right instances
-       }
-       deriving( Eq, Ord, Show, Generic, NFData, Binary )
-
--- | A closed proto rule lists its original rule modulo E, the corresponding
--- variant(s) modulo AC, and if required the assertion soundness proof.
--- When using auto-sources, all non-trivial variants of a ClosedProtoRule are
--- split up into multiple ClosedProtoRules. Auto-sources also only adds
--- actions only to closed rules. Opening such rules keeps the AC rules s.t.
--- they can be exported.
-data ClosedProtoRule = ClosedProtoRule
-       { _cprRuleE         :: ProtoRuleE      -- original rule modulo E
-       , _cprRuleAC        :: ProtoRuleAC     -- variant(s) modulo AC
-       }
-       deriving( Eq, Ord, Show, Generic, NFData, Binary )
-
-type OpenRuleCache = [IntrRuleAC]
-
-data ClosedRuleCache = ClosedRuleCache
-       { _crcRules               :: ClassifiedRules
-       , _crcRawSources          :: [Source]
-       , _crcRefinedSources      :: [Source]
-       , _crcInjectiveFactInsts  :: S.Set FactTag
-       }
-       deriving( Eq, Ord, Show, Generic, NFData, Binary )
-
-$(mkLabels [''OpenProtoRule, ''DiffProtoRule, ''ClosedProtoRule, ''ClosedRuleCache])
-
-instance HasRuleName OpenProtoRule where
-    ruleName = ruleName . L.get oprRuleE
-
-instance HasRuleName DiffProtoRule where
-    ruleName = ruleName . L.get dprRule
-
-instance HasRuleName ClosedProtoRule where
-    ruleName = ruleName . L.get cprRuleAC
-
-
--- | Get an OpenProtoRule's name
-getOpenProtoRuleName :: OpenProtoRule -> String
-getOpenProtoRuleName (OpenProtoRule ruE _) = getRuleName ruE
-
--- | Add the diff label to an OpenProtoRule
-addProtoDiffLabel :: OpenProtoRule -> String -> OpenProtoRule
-addProtoDiffLabel (OpenProtoRule ruE ruAC) label = OpenProtoRule (addDiffLabel ruE label) (fmap ((flip addDiffLabel) label) ruAC)
-
-equalOpenRuleUpToDiffAnnotation :: OpenProtoRule -> OpenProtoRule -> Bool
-equalOpenRuleUpToDiffAnnotation (OpenProtoRule ruE1 ruAC1) (OpenProtoRule ruE2 ruAC2) =
-  equalRuleUpToDiffAnnotationSym ruE1 ruE2 && length ruAC1 == length ruAC2 &&
-  all (uncurry equalRuleUpToDiffAnnotationSym) (zip ruAC1 ruAC2)
-
--- Relation between open and closed rule sets
----------------------------------------------
-
--- | All intruder rules of a set of classified rules.
-intruderRules :: ClassifiedRules -> [IntrRuleAC]
-intruderRules rules = do
-    Rule (IntrInfo i) ps cs as nvs <- joinAllRules rules
-    return $ Rule i ps cs as nvs
-
--- | Open a rule cache. Variants and precomputed case distinctions are dropped.
-openRuleCache :: ClosedRuleCache -> OpenRuleCache
-openRuleCache = intruderRules . L.get crcRules
-
--- | Open a protocol rule; i.e., drop variants and proof annotations.
-openProtoRule :: ClosedProtoRule -> OpenProtoRule
-openProtoRule r = OpenProtoRule ruleE ruleAC
-  where
-    ruleE   = L.get cprRuleE r
-    ruleAC' = L.get cprRuleAC r
-    ruleAC  = if equalUpToTerms ruleAC' ruleE
-               then []
-               else [ruleAC']
-
--- | Unfold rule variants, i.e., return one ClosedProtoRule for each
--- variant
-unfoldRuleVariants :: ClosedProtoRule -> [ClosedProtoRule]
-unfoldRuleVariants (ClosedProtoRule ruE ruAC@(Rule ruACInfoOld ps cs as nvs))
-   | isTrivialProtoVariantAC ruAC ruE = [ClosedProtoRule ruE ruAC]
-   | otherwise = map toClosedProtoRule variants
-        where
-          ruACInfo i = ProtoRuleACInfo (rName i (L.get pracName ruACInfoOld)) rAttributes (Disj [emptySubstVFresh]) loopBreakers
-          rAttributes = L.get pracAttributes ruACInfoOld
-          loopBreakers = L.get pracLoopBreakers ruACInfoOld
-          rName i oldName = case oldName of
-            FreshRule -> FreshRule
-            StandRule s -> StandRule $ s ++ "___VARIANT_" ++ show i
-
-          toClosedProtoRule (i, (ps', cs', as', nvs'))
-            = ClosedProtoRule ruE (Rule (ruACInfo i) ps' cs' as' nvs')
-          variants = zip [1::Int ..] $ map (\x -> apply x (ps, cs, as, nvs)) $ substs (L.get pracVariants ruACInfoOld)
-          substs (Disj s) = map (`freshToFreeAvoiding` ruAC) s
-
--- | Close a protocol rule; i.e., compute AC variant and source assertion
--- soundness sequent, if required.
-closeProtoRule :: MaudeHandle -> OpenProtoRule -> [ClosedProtoRule]
-closeProtoRule hnd (OpenProtoRule ruE [])   = [ClosedProtoRule ruE (variantsProtoRule hnd ruE)]
-closeProtoRule _   (OpenProtoRule ruE ruAC) = map (ClosedProtoRule ruE) ruAC
-
--- | Close an intruder rule; i.e., compute maximum number of consecutive applications and variants
---   Should be parallelized like the variant computation for protocol rules (JD)
-closeIntrRule :: MaudeHandle -> IntrRuleAC -> [IntrRuleAC]
-closeIntrRule hnd (Rule (DestrRule name (-1) subterm constant) prems@((Fact KDFact _ [t]):_) concs@[Fact KDFact _ [rhs]] acts nvs) =
-  if subterm then [ru] else variantsIntruder hnd id False ru
-    where
-      ru = (Rule (DestrRule name (if runMaude (unifiableLNTerms rhs t)
-                              then (length (positions t)) - (if (isPrivateFunction t) then 1 else 2)
-                              -- We do not need to count t itself, hence - 1.
-                              -- If t is a private function symbol we need to permit one more rule
-                              -- application as there is no associated constructor.
-                              else 0) subterm constant) prems concs acts nvs)
-        where
-           runMaude = (`runReader` hnd)
-closeIntrRule hnd ir@(Rule (DestrRule _ _ False _) _ _ _ _) = variantsIntruder hnd id False ir
-closeIntrRule _   ir                                        = [ir]
-
-
--- | Close a rule cache. Hower, note that the
--- requires case distinctions are not computed here.
-closeRuleCache :: [LNGuarded]        -- ^ Restrictions to use.
-               -> [LNGuarded]        -- ^ Source lemmas to use.
-               -> SignatureWithMaude -- ^ Signature of theory.
-               -> [ClosedProtoRule]  -- ^ Protocol rules with variants.
-               -> OpenRuleCache      -- ^ Intruder rules modulo AC.
-               -> Bool               -- ^ Diff or not
-               -> ClosedRuleCache    -- ^ Cached rules and case distinctions.
-closeRuleCache restrictions typAsms sig protoRules intrRules isdiff = -- trace ("closeRuleCache: " ++ show classifiedRules) $
-    ClosedRuleCache
-        classifiedRules rawSources refinedSources injFactInstances
-  where
-    ctxt0 = ProofContext
-        sig classifiedRules injFactInstances RawSource [] AvoidInduction Nothing
-        (error "closeRuleCache: trace quantifier should not matter here")
-        (error "closeRuleCache: lemma name should not matter here") [] isdiff
-        (all isSubtermRule {-- $ trace (show destr ++ " - " ++ show (map isSubtermRule destr))-} destr) (any isConstantRule destr)
-
-    -- inj fact instances
-    injFactInstances =
-        simpleInjectiveFactInstances $ L.get cprRuleE <$> protoRules
-
-    -- precomputing the case distinctions: we make sure to only add safety
-    -- restrictions. Otherwise, it wouldn't be sound to use the precomputed case
-    -- distinctions for properties proven using induction.
-    safetyRestrictions = filter isSafetyFormula restrictions
-    rawSources         = precomputeSources ctxt0 safetyRestrictions
-    refinedSources     = refineWithSourceAsms typAsms ctxt0 rawSources
-
-    -- Maude handle
-    hnd = L.get sigmMaudeHandle sig
-
-    -- close intruder rules
-    intrRulesAC = concat $ map (closeIntrRule hnd) intrRules
-
-    -- classifying the rules
-    rulesAC = (fmap IntrInfo                      <$> intrRulesAC) <|>
-              ((fmap ProtoInfo . L.get cprRuleAC) <$> protoRules)
-
-    anyOf ps = partition (\x -> any ($ x) ps)
-
-    (nonProto, proto) = anyOf [isDestrRule, isConstrRule] rulesAC
-    (constr, destr)   = anyOf [isConstrRule] nonProto
-
-    -- and sort them into ClassifiedRules datastructure for later use in proofs
-    classifiedRules = ClassifiedRules
-      { _crConstruct  = constr
-      , _crDestruct   = destr
-      , _crProtocol   = proto
-      }
-
-
--- | Returns true if the REFINED sources contain open chains.
-containsPartialDeconstructions :: ClosedRuleCache    -- ^ Cached rules and case distinctions.
-                     -> Bool               -- ^ Result
-containsPartialDeconstructions (ClosedRuleCache _ _ cases _) =
-      sum (map (sum . unsolvedChainConstraints) cases) /= 0
-
--- | Add an action to a closed Proto Rule.
---   Note that we only add the action to the variants modulo AC, not the initial rule.
-addActionClosedProtoRule :: ClosedProtoRule -> LNFact -> ClosedProtoRule
-addActionClosedProtoRule (ClosedProtoRule e ac) f
-   = ClosedProtoRule e (addAction ac f)
 
 -- | Open theories can be extended. Invariants:
 --   1. Lemma names are unique.
